Split up Overtime.tsx
