+++ src/factories/index.ts
@@ -1,6 +1,7 @@
 import { register } from "fishery"
 import analyticsContext from "./analyticsContext"
 import configurationContext from "./configurationContext"
+import dayLogData from "./dayLogData"
 import exposureContext from "./exposureContext"
 import exposureDatum from "./exposureDatum"
 import gaenStrategy from "./gaenStrategy"
@@ -10,6 +11,7 @@ import rawExposure from "./rawExposure"
 export const factories = register({
   analyticsContext,
   configurationContext,
+  dayLogData,
   exposureContext,
   exposureDatum,
   gaenStrategy,
